+++ src/plone/app/mosaic/registry.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 from copy import deepcopy
-from operator import itemgetter
 from plone.app.mosaic.interfaces import IMosaicRegistryAdapter
 from plone.app.mosaic.utils import extractFieldInformation
 from plone.dexterity.utils import iterSchemataForType
@@ -41,6 +40,11 @@ def getCategoryIndex(tiles, category):
 
 def weightedSort(x):
     return x[1]['weight']
+
+
+def safe_weight_sortkey(x):
+    weight = x.get('weight', None)
+    return weight if weight is not None else 9000
 
 
 @implementer(IMosaicRegistryAdapter)
@@ -89,7 +93,7 @@ class MosaicRegistry(object):
                     items = list(items.values())
                 if items:
                     action['items'] = items
-                    action['items'].sort(key=itemgetter('weight'))
+                    action['items'].sort(key=safe_weight_sortkey)
                     for x in action['items']:
                         x['value'] = x['name']
 
@@ -144,7 +148,7 @@ class MosaicRegistry(object):
                 config['formats'][index]['actions'].append(format)
         # sort the formats
         for format in config['formats']:
-            format['actions'].sort(key=itemgetter('weight'))
+            format['actions'].sort(key=safe_weight_sortkey)
         return config
 
     def mapTinyMCEActionCategories(self, settings, config):
@@ -168,7 +172,7 @@ class MosaicRegistry(object):
             if index is not None:
                 config['richtext_toolbar'][index]['actions'].append(action)
         for group in config['richtext_toolbar']:
-            group['actions'].sort(key=itemgetter('weight'))
+            group['actions'].sort(key=safe_weight_sortkey)
         return config
 
     def mapTinyMCEContextMenuFormats(self, settings, config):
@@ -179,7 +183,7 @@ class MosaicRegistry(object):
             if index is not None:
                 config['richtext_contextmenu'][index]['actions'].append(action)
         for group in config['richtext_contextmenu']:
-            group['actions'].sort(key=itemgetter('weight'))
+            group['actions'].sort(key=safe_weight_sortkey)
         return config
 
     # def mapStructureTiles(self, settings, config):
@@ -193,7 +197,7 @@ class MosaicRegistry(object):
     #        if index is not None:
     #            config['tiles'][index]['tiles'].append(tile)
     #    for tile in config['tiles']:
-    #        tile['tiles'].sort(key=itemgetter('weight'))
+    #        tile['tiles'].sort(key=safe_weight_sortkey)
     #    return config
     #
     # def mapApplicationTiles(self, settings, config):
@@ -205,7 +209,7 @@ class MosaicRegistry(object):
     #        if index is not None:
     #            config['tiles'][index]['tiles'].append(tile)
     #    for tile in config['tiles']:
-    #        tile['tiles'].sort(key=itemgetter('weight'))
+    #        tile['tiles'].sort(key=safe_weight_sortkey)
     #    return config
 
     def mapTiles(self, settings, config, tile_category):
@@ -218,7 +222,7 @@ class MosaicRegistry(object):
             if index is not None:
                 config['tiles'][index]['tiles'].append(tile)
         for tile in config['tiles']:
-            tile['tiles'].sort(key=itemgetter('weight'))
+            tile['tiles'].sort(key=safe_weight_sortkey)
         return config
 
     # BBB: needs a bit of thought, I'm nowhere near satisfied with this